pyproject.toml
[tool.poetry]
name = "magic-ball-bot"
version = "0.1.0"
description = ""
authors = ["ribeyra <ribeyra@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "13.15"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
